Rename xhr extension file.
