+++ src/util/extend-xhr.js
